requirements.txt
OpenCV2-Python
numpy
tensorflow
glob